pyproject.toml
[tool.poetry]
name = "currency"
version = "0.1.0"
description = ""
authors = ["spirin-vadim <vadis99@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
